+++ NEWS.md
@@ -1,7 +1,8 @@
 # squid 0.1.0.9000
 
+* Change module's name: from "Random slope" to "Random regression".
 
 
 # squid 0.1.0
 
-First release.
+* First release.